Facebook Authorisation Works to an Extent
The next thing I’ll work on is saving the user in the database after
authentication with Facebook and creating that ‘Facebooker’ column to
remember if they’d signed up with Facebook.

This way I’m able to use PHP to determine whether we should use Ion
auth or Facebook’s SDK to authenticate users.